unification of hiding logic
