--- Explorer/Assets/DCL/AvatarRendering/AvatarShape/Helpers/AvatarWearableHide.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 855e46ddf1ab4037ad1e97a5fee2146e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
